feat: added getUserPref() returns num array fix: changeUserPref not deleting properly

+++ Database/mysqlconnect.php
@@ -184,14 +184,17 @@ class mysqlConnect {
 	//user funcitons
 	public function changeUserPref($username, $pref_array) {
 		$uid = getUIDbyUsername($username, $this->mydb);
-		$formatted_labels = "'" . implode("','", $pref_arr) . "'";
+		if($uid === null) {
+			return array('status' => 'Error');
+		}
+		$formatted_labels = "'" . implode("','", $pref_array) . "'";
 
 		$checkNumQuery = "SELECT COUNT(uid) AS total FROM user_pref WHERE uid = ".$uid.";" ;
 		$response = handleQuery($checkNumQuery, $this->mydb, "Query Status: User Pref Count Successfull");
 		$sum_arr = $response->fetch_assoc();
 		if ($sum_arr['total'] > 0) {
 			$delete_query = "DELETE FROM user_pref WHERE uid = ".$uid.";" ;
-			$response = handleQuery($checkNumQuery, $this->mydb, "Query Status: User Pref Delete Successfull");
+			$response = handleQuery($delete_query, $this->mydb, "Query Status: User Pref Delete Successfull");
 			if ($response === false) {
 				return array('status' => 'Error');
 			}
@@ -200,8 +203,10 @@ class mysqlConnect {
 		$query = "INSERT INTO user_pref (uid, label_id)
 		SELECT ".$uid.", label_id
 		FROM labels WHERE label_name IN (".$formatted_labels.");";
-		$response = handleQuery($checkNumQuery, $this->mydb, "Query Status: Add User: ".$username." Prefs Successfull");
-		
+		$response = handleQuery($query, $this->mydb, "Query Status: Add User: ".$username." Prefs Successfull");
+
+		//$debug = getUserPref($username, $this->mydb);
+		//print_r($debug);
 		return array('status' => $response ? 'Success' : 'Error');
 	}
 
@@ -225,6 +230,12 @@ function showTwoAR ($array) {
 	foreach($array as $row) {
 		showAr($row);
 		echo "\n";
+	}
+}
+
+function showRows($array) {
+	foreach($array as $row) {
+		echo $row."\n";
 	}
 }
 
@@ -329,6 +340,10 @@ $chickenRecipes = [
 
 showTwoAr($testObj->checkRecipe('Chicken', 'high-protein, dairy-free'));
 showTwoAr($testObj->checkRecipe('Caesar Salad'));
+
+$test_prefs = ['dairy-free', 'gluten-free', 'high-protein', 'Kosher'];
+$testObj->changeUserPref('Bob', $test_prefs);
+
 
 /*showAr($testObj->registerAccount("Bob","bobby@gmail.com", "crabcake"));
 showAr($testObj->registerAccount("dummyuser","dummy@email.com", "dummypass"));
